Share OpenAI header builder (oh-tab-4xv)
